fix cart

+++ 22_WebBanXe(Eshop)/src/vehicle_web/cart/views.py
@@ -1,23 +1,40 @@
-from django.shortcuts import render, redirect
-from .models import CartItem
-from vehicle.models import Vehicle
+from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+from .models import Cart
 
-def view_cart(request):
-	cart_items = CartItem.objects.filter(user=request.user)
-	total_price = sum(item.vehicle.price * item.quantity for item in cart_items)
-	return render(request, 'cart/cart.html', {'cart_items': cart_items, 'total_price': total_price})
+@login_required
+def add_to_cart(request, product_id):
+    cart_item = Cart.objects.filter(user=request.user, product=product_id).first()
 
-def add_to_cart(request, vehicle_id):
-    if request.method == 'POST':
-        vehicle = Vehicle.objects.get(id=vehicle_id)
-        cart_item, created = CartItem.objects.get_or_create(vehicle=vehicle, user=request.user)
+    if cart_item:
         cart_item.quantity += 1
         cart_item.save()
-        return redirect('cart:view_cart')
+        messages.success(request, "Item added to your cart.")
     else:
-        return HttpResponseNotAllowed(['POST'])
+        Cart.objects.create(user=request.user, product=product_id)
+        messages.success(request, "Item added to your cart.")
 
-def remove_from_cart(request, item_id):
-	cart_item = CartItem.objects.get(id=item_id)
-	cart_item.delete()
-	return redirect('cart:view_cart')
+    return redirect("cart_detail")
+
+@login_required
+def remove_from_cart(request, cart_item_id):
+    cart_item = get_object_or_404(Cart, id=cart_item_id)
+
+    if cart_item.user == request.user:
+        cart_item.delete()
+        messages.success(request, "Item removed from your cart.")
+
+    return redirect("cart_detail")
+
+@login_required
+def cart_detail(request):
+    cart_items = Cart.objects.filter(user=request.user)
+    total_price = sum(item.quantity * item.product.price for item in cart_items)
+
+    context = {
+        "cart_items": cart_items,
+        "total_price": total_price,
+    }
+
+    return render(request, "cart/cart_detail.html", context)
